Add Stable-water analysis to OpenMMDL analysis & fix
If no Ligand is present during the simulation, the option to add analysis resulted in a crash before.

Now, it performs analysis of the water during the MD simulation to find unbiased stable water molecules. A PDB with representative water molecules is written as well as each cluster that was found in a seperate PDB containing all cluster members.

Default is set to 75 % of the MD and clustering on 1Angstrom distance. However this can be adjusted within the script

TODO:

-allow the user to choose during simulation-setup if analysis of water shall be performed, also if a ligand was present during the simulation.
- allow for command line input to adjust the clustering of the watermoelcules differently
- createthe output directory not at the user's current location if the analysis is run standalone